Allow connection to bitbox-base electrs
This commit adds support to connect with electrs running testnet on the
BitBox Base. A new postConnectElectrumHandler was created for this
purpose. It takes a bitboxBaseID, from that gets the connected bitbox
Bases'IP and initiates a connection with the base.

The electrs port and network type (mainnet/testnet) are fetched with the
GetEnv get request.

Once connected, other active backends will be unset, such that
only bitcoin accounts will be available. This needs further tweaking,
such that users cannot re-active these de-activated accounts in the app
settings, as long as the connection with a bitbox base is active.