Use Protobuf for data serialization with the base
Data that is communicated between the base and the desktop app is now
serialized with protobuf. Implment the serialization for the nodeinfo
and systemenv calls.
Instead of a generic interface, the events between the middleware and
handlers packages are now expected to be protbuf messages.
Close websocket connections properly with callbacks and channel events.
When a base disconnects, its connection is now properly deconstructed.